requirements.txt
python-dotenv
simpleaudio